pyproject.toml
[tool.poetry]
name = "django-debug-toolbar-user-switcher"
version = "2.0.3"
description = "Panel for the Django Debug toolbar to quickly switch between users."
authors = ["Thread Engineering <tech@thread.com>"]
license = "BSD-3-Clause"
include = ["debug_toolbar_user_switcher/templates/debug_toolbar_user_switcher/*"]
packages = [
    { include = "debug_toolbar_user_switcher" },
]

[tool.poetry.dependencies]
python = "^3.6"
django = ">=2.2,<5"
django-debug-toolbar = ">=2,<4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
